Moved _section/ForPartner to seperate folder under for-partner. Renamed components. Fix copywrite for fpfeatureslider.
