+++ src/app/for-partners/_sections/Benefits.tsx
@@ -103,8 +103,8 @@ const Benefits = () => {
         <path
           d="M1510 2137.56C1308.89 2291.03 1167.68 1479.09 886.104 1688.05C464.13 2001.21 23.1114 1775.88 397.856 1516.33C784.185 1248.76 1686.93 974.095 1454.5 1329C1330.62 1518.15 663.065 1290.64 752 1056.5C850 798.5 622.321 736.5 342.637 1016.72C56.7074 1303.2 -344.5 1027.5 342.637 809.5C1225.25 529.482 1436 231 1510 328C1648.82 509.963 149.499 528.5 -25.0006 7.5"
           stroke="#E97958"
-          stroke-width="14.2383"
-          stroke-linecap="round"
+          strokeWidth="14.2383"
+          strokeLinecap="round"
         />
       </svg>
