+++ src/app/for-partners/_sections/Hero.tsx
