+++ src/app/for-partners/_sections/PartnerSteps.tsx
@@ -110,7 +110,7 @@ const STEPS = [
 //   )
 // }
 
-const PathScrollSteps = () => {
+const PartnerSteps = () => {
   const sectionRef = useRef<HTMLDivElement | null>(null)
   const svgRef = useRef<SVGSVGElement | null>(null)
   const stepRefs = useRef<HTMLDivElement[]>([])
@@ -221,8 +221,8 @@ const PathScrollSteps = () => {
         <path
           d="M-51.0009 25.8586C149.08 -99.3193 323.866 454.935 603.998 284.5C1023.82 29.0787 1560.83 275.802 1188 487.5C803.645 705.742 115.067 988.259 147.499 687C151.976 645.42 910.015 537.49 835.499 812.5C760.984 1087.51 1501 559.575 1501 710.5C1501 753 1570.5 899 905.499 1070.5C676.482 1129.56 -198.425 1402.98 -93.001 1174C31.9986 902.5 1155.5 1317 1470 1558.5"
           stroke="#E97958"
-          stroke-width="14.2383"
-          stroke-linecap="round"
+          strokeWidth="14.2383"
+          strokeLinecap="round"
         />
       </svg>
 
@@ -247,4 +247,4 @@ const PathScrollSteps = () => {
   )
 }
 
-export default PathScrollSteps
+export default PartnerSteps
